refactor: move commit-producing APIs to new file
We're going to have a bunch of methods on `ConversationGuard` eventually,
so we may as well set them up in a new file to establish a pattern now.
